refactor(gsd): extract migrate execution service
